[BE0059] Add Constraint to Fields in DB

+++ src/main/java/com/smart/tailor/entities/LaborQuantity.java
@@ -27,9 +27,9 @@ public class LaborQuantity extends AuditEntity implements Serializable {
 
     private Integer laborQuantityMaxQuantity;
 
-    private Double laborQuantityMinPrice;
+    private Integer laborQuantityMinPrice;
 
-    private Double laborQuantityMaxPrice;
+    private Integer laborQuantityMaxPrice;
 
     private Boolean status;
 }
